Warning: Non-*ABI* compatible commit. You must rebuild KDE plugins after updating or you might/will experience crashes. API compatible.
This commit:

1) Adds the ability for applications using attica as a library to pass
flags to the provider manager, notably to disable plugin loading so that
deterministic QtPlatformDependent behavior can be used.

2) Updates the PlatformDependent API to add a setNam call to allow
setting a custom QNetworkAccessManager.

3) Implements setting/getting QNAMs in the ProviderManager in a
thread-safe way. QNetworkAccessManager is not thread-safe and so using
the same QNAM across multiple threads can result in crashes if one of
them is deleted, most notably because it takes ownership of a
QNetworkProxyFactory object so simply passing in a proxyfactory is not
good enough.