Accept any HTTP status between 100 and 199 as benign
In December 2017, a new status code (103) was added to the HTTP
standard, and as one prominent OCS provider now uses this status
code, Attica needs to be able to handle it more gracefully than
just reporting an unknown OCS error (which is reported to the
user, but has no actual effect on things). As it is a benign
status, we simply accept it and forward it to the consumer as
being not an error. Specifically, we do this by changing from
accepting only status 100 as an informational status, to
accepting any informational status (since any actual errors are
reported through other status codes).

Differential Revision: https://phabricator.kde.org/D20077