Set a proper agent string for Attica requests
Summary:
Rather than simply accepting the default value for the agent string,
which suggests that this library is Mozilla, we want to make requests
specifically with the name of what's actually making the request.

The usual reason for using the Mozilla identifier is that it makes
some websites send out webpages differently, but as we are interfacing
with a different type of server, we want to make sure that we are
sending out who we really are, rather than spoofing.

Reviewers: #frameworks, #knewstuff, #plasma, starbuck, ronaldv, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25440