Add a default provider (as per documentation)
Summary:
Prior to this patch, without the KDE Platform Dependent, Attica
would fail to correctly initialise providers from KNewStuff which
were made according to the documentation in KNewStuff when run
in systems which did not have Plasma Desktop installed (which is
what provides the above mentioned platformdependent). With this
patch, while saving and restoring such information doesn't work,
at least they will work as intended.

This removes the need for including the provider URL in our own
released knsrc files, when it is explicitly suggested that it
should not be required.

Note: The documented method will work just fine for any
Plasma Desktop provided ones already, as any Plasma install
will have the correct plugin, but for any other file, this really
should still work the way the documentation says it is
supposed to.

See this email for some further context:
https://mail.kde.org/pipermail/kde-frameworks-devel/2020-January/102643.html

Reviewers: kossebau, #frameworks, #knewstuff, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27123