Honour the job being aborted immediately
Occasionally, a job will be started and then immediately cancelled. This
can cause the job to end up still attempting to run while it is in the
process of being destructed, which in turn leads to accessing bits of
data which are no longer available, and in the end a crash. This fixes
that problem by ensuring we actually honour abort calls internally as
well.

Also adds a little test for listing content, and for switching the sort order.

BUG:429939

FIXED-IN:5.78