Add http headers to Metadata (and set it from BaseJob)
This allows an api consumer to read various useful information from a
job in the case of an error occurring. As a specific example, if we get
a 503 status, a consumer may wish to read out the Retry-After header to
launch the job again after that time (or give the user some useful
information regarding this situation).