Longer timeout for the provider test
While this is something of an issue, it is not an issue we have control
over from within attica itself. In short: We are not testing the server
that we are running this test against, we are testing the functionality
of our provider classes, and having a very short timeout causes false
positives for this test. So, longer timeout and we should have a more
generally happy CI.